Add new meta-implementation of base-e logarithm
Unified denormal and large exponent inputs handling for the logarithm
tries to make use of integer arithmetics as much as possible. This
special handling is necessary when using FastReciprocal because 1/x may
either overflow to +infty or flush to 0, which is very bad for the
polynomial evaluation step.

For a faster 1/x approximation without FastReciprocal available, we
should use an ML_Table instead of a Division operation (high latency in
general, and an accuracy overkill as only the first bits are used to
address the table).

Use fast CGPE scheme for the vectorized-logarithm