Increase 'points' value if fpminimax returns error
If sollya_poly is a Sollya error object, try doubling the Sollya
'points' variable until dirtyfindzeros can find all zeros in the
interval, or until 'points' is >= 10000, which is an arbitrary value
such that the overhead of using more sampling points is not too much.
With such a value, the overhead should remain below a 10x factor. See
Sollya 'time' and 'timing' help for timing operations within Sollya
itself.

Another option would be to relax the approx_interval by increasing the
upper bound (say, for example, to 1/1000 of its absolute value).
Increasing the upper bound can make the interval asymmetric (if the
interval was centered on 0 before), helping in finding zeros of the
approximation error function.

See the discussion about this with Sylvain Chevillard on the
sollya-users mailing list for a better insight about Sollya's fpminimax
returning error when using an incomplete basis of monomials on an
interval containing zero.