Catch a TypeError that could make ml-debug not break
When a silly meta-function developer (namely, I) would recursively
create a Constant() whose value was already a Constant, Metalibm would
crash even with --ml-debug enabled, whereas it was expected to break on
error.  Maybe we want to catch more exceptions as well, but I guess this
should be discussed before any other change.