Cleaning gappa support testing
Gappa has changed the way it displays its help message. This means
that when metalibm is testing whether gappa is installed or not, the help
get dumped on the stdout. We are now redirecting all gappa output to /dev/null
and simply testing the return call of calling gappa in a subprocess call.